Channel routes cleared and doouble checked

+++ app/api/channel_routes.py
@@ -14,28 +14,29 @@ channel_routes = Blueprint('channel', __name__, url_prefix="/channels")
 @channel_routes.route('/', methods=["POST"])
 def create_channel():
     # user = User.query.get(userId)
-        params = request.get_json()
-        server_id = params['server_id']
-        form = ChannelForm()
-        form['csrf_token'].data = request.cookies['csrf_token']
-        if form.validate_on_submit():
-            channel = Channel(name=form.data['name'], server_id=server_id)
-            db.session.add(channel)
-            db.session.commit()
-            return channel.to_dict(), 201
-        else:
-            return {'errors': error_messages(form.errors)}, 401
+    params = request.get_json()
+    server_id = params['server_id']
+    form = ChannelForm()
+    form['csrf_token'].data = request.cookies['csrf_token']
+    if form.validate_on_submit():
+        channel = Channel(name=form.data['name'], server_id=server_id)
+        db.session.add(channel)
+        db.session.commit()
+        return channel.to_dict(), 201
+    else:
+        return {'errors': error_messages(form.errors)}, 401
 
 # TODO ——————————————————————————————————————————————————————————————————————————————————
 # *                                   READ
 # TODO ——————————————————————————————————————————————————————————————————————————————————
 
-# The reason we include a server_id parameter, is so that we only get the channels that belong to a specific server. 
+# The reason we include a server_id parameter,
+# is so that we only get the channels that belong to a specific server.
+
+
 @channel_routes.route('/<int:server_id>', methods=["GET"])
 def all_channels(server_id):
-    # * This query returns a non-Pythonic list of all channels
     channels = Channel.query.filter_by(server_id=server_id)
-    # * This returns a key/val pair of channels in JSON format
     return {'channels': [channel.to_dict() for channel in channels]}
 
 # TODO ——————————————————————————————————————————————————————————————————————————————————
@@ -70,7 +71,13 @@ def update_channel(channelId):
 
 @channel_routes.route('/<int:channel_id>', methods=['DELETE'])
 def delete_channel(channelId):
+
     channel = Channel.query.get(channelId)
-    db.session.delete(channel)
-    db.session.commit()
+    if channel:
+        server_id = channel.server_id
+        server = Server.query.get(server_id)
+        if server.owner_id == current_user.id:
+            db.session.delete(channel)
+            db.session.commit()
+            return channel.to_dict()
     return f"channel {channelId} has been deleted!"
